[🚀] (Heap Extract)

+++ 0x08-palindrome_integer/palindrome.h
@@ -1,7 +1,7 @@
 #ifndef _PALINDROME_H
 #define _PALINDROME_H
 
-int palindrome(unsigned long n, unsigned long initial, int base, int lastBase);
+int palindrome(unsigned long n, unsigned long initial, unsigned long base, unsigned long lastBase);
 
 #define is_palindrome(N) palindrome((N), (N), 0, 0)
